refactor: move clearLocalStorage function to utils and simplify localStorage clearing logic
